pyproject.toml
[project]
name = "test-agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.4",
    "pydantic>=2.10.6",
    "python-docx>=1.1.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]
